LabelSet.String(): use different implementation for go 1.20
