+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/prometheus/common
 
-go 1.21
+go 1.20
 
 require (
 	github.com/alecthomas/kingpin/v2 v2.4.0
